Free CoreGraphics memory allocated in Darwin implementation
Before this patch, CaptureRect was passing the return values of
multiple CoreGraphics methods into each other directly. The
return values of two of these methods, CGImageGetDataProvider and
CGDisplayCreateImage, are required to be released by the caller
per documentation. Because the return values were not even being
captured, they could not be released, resulting in a significant
memory leak.

This patch stores those return values separately and frees them
before returning the captured image.